requirements.txt
discord.py
python-dotenv
rapidfuzz